+++ client/pkg/circuitbreaker/circuit_breaker_test.go
